+++ config.py
@@ -22,7 +22,8 @@ PLUGIN_ROOT = os.path.dirname(os.path.realpath(__file__))
 WEB_DIRECTORY = os.path.join(PLUGIN_ROOT, "web")
 JAVASCRIPT_PATH = os.path.join(WEB_DIRECTORY, "js")
 CSS_PATH = os.path.join(WEB_DIRECTORY, "css")
-PLACEHOLDER_IMAGE_PATH = os.path.join(WEB_DIRECTORY, "/images/placeholder.jpeg")
+# Corrected path construction to avoid issues with leading slashes
+PLACEHOLDER_IMAGE_PATH = os.path.join(WEB_DIRECTORY, "images", "placeholder.jpeg")
 
 # Get ComfyUI directories using folder_paths
 COMFYUI_ROOT_DIR = folder_paths.base_path
