+++ server/__init__.py
@@ -1,0 +1,5 @@
+# ================================================
+# File: server/__init__.py
+# ================================================
+# Import the routes package to ensure all route decorators are executed.
+from . import routes
