+++ web/js/civitaiDownloader.js
@@ -6,6 +6,7 @@ import { Feedback } from "./ui/feedback.js";
 import { renderSearchResults as renderSearchResultsExternal } from "./ui/searchRenderer.js";
 import { renderDownloadList as renderDownloadListExternal } from "./ui/statusRenderer.js";
 import { renderDownloadPreview as renderDownloadPreviewExternal } from "./ui/previewRenderer.js";
+import { modalTemplate } from "./ui/templates.js";
 
 console.log("Loading Civicomfy UI...");
 
@@ -215,176 +216,8 @@ class CivitaiDownloaderUI {
         this.modal = document.createElement('div');
         this.modal.className = 'civitai-downloader-modal';
         this.modal.id = 'civitai-downloader-modal';
-        // HTML structure including Settings Tab
-        this.modal.innerHTML = `
-            <div class="civitai-downloader-modal-content">
-                <div class="civitai-downloader-header">
-                    <h2>Civicomfy</h2>
-                    <button class="civitai-close-button" id="civitai-close-modal">&times;</button>
-                </div>
-                <div class="civitai-downloader-body">
-                    <div class="civitai-downloader-tabs">
-                        <button class="civitai-downloader-tab active" data-tab="download">Download</button>
-                        <button class="civitai-downloader-tab" data-tab="search">Search</button>
-                        <button class="civitai-downloader-tab" data-tab="status">Status <span id="civitai-status-indicator" style="display: none;">(<span id="civitai-active-count">0</span>)</span></button>
-                        <button class="civitai-downloader-tab" data-tab="settings">Settings</button>
-                    </div>
-                    <div id="civitai-tab-download" class="civitai-downloader-tab-content active">
-                        <form id="civitai-download-form">
-                            <div class="civitai-form-group">
-                                <label for="civitai-model-url">Model URL or ID</label>
-                                <input type="text" id="civitai-model-url" class="civitai-input" placeholder="e.g., https://civitai.com/models/12345 or 12345" required>
-                            </div>
-                            <p style="font-size: 0.9em; color: #ccc; margin-top: -10px; margin-bottom: 15px;">You can optionally specify a version ID using "?modelVersionId=xxxxx" in the URL or in the field below.</p>
-                            <div class="civitai-form-row">
-                                <div class="civitai-form-group">
-                                    <label for="civitai-model-type">Model Type (Save Location)</label>
-                                    <select id="civitai-model-type" class="civitai-select" required></select>
-                                </div>
-                                <div class="civitai-form-group">
-                                    <label for="civitai-model-version-id">Version ID (Optional)</label>
-                                    <input type="number" id="civitai-model-version-id" class="civitai-input" placeholder="Overrides URL/Latest">
-                                </div>
-                            </div>
-                             <div class="civitai-form-row">
-                                <div class="civitai-form-group">
-                                    <label for="civitai-custom-filename">Custom Filename (Optional)</label>
-                                    <input type="text" id="civitai-custom-filename" class="civitai-input" placeholder="Leave blank to use original name">
-                                </div>
-                                <div class="civitai-form-group">
-                                    <label for="civitai-connections">Connections</label>
-                                    <input type="number" id="civitai-connections" class="civitai-input" value="${this.settings.numConnections}" min="1" max="16" step="1" required disabled>
-                                    <p style="font-size: 0.9em; color: #ccc; margin-top: 7px; margin-bottom: 15px;">Disabled: Only single connection possible for now</p>
-                                    </div>
-                            </div>
-                            <div class="civitai-form-group inline">
-                                 <input type="checkbox" id="civitai-force-redownload" class="civitai-checkbox">
-                                <label for="civitai-force-redownload">Force Re-download (if exists)</label>
-                            </div>
-
-                            <div id="civitai-download-preview-area" class="civitai-download-preview-area" style="margin-top: 25px; margin-bottom: 25px; padding-top: 15px; border-top: 1px solid var(--border-color, #444);">
-                            <!-- Preview content will be injected here -->
-                            </div>
-
-                            <button type="submit" id="civitai-download-submit" class="civitai-button primary">Start Download</button>
-                        </form>
-                    </div>
-                    <div id="civitai-tab-search" class="civitai-downloader-tab-content">
-                        <form id="civitai-search-form">
-                            <div class="civitai-search-controls">
-                                <input type="text" id="civitai-search-query" class="civitai-input" placeholder="Search Civitai..."> <!-- Remove required -->
-                                <select id="civitai-search-type" class="civitai-select">
-                                    <option value="any">Any Type</option>
-                                    <!-- Model types populated here -->
-                                </select>
-                                <!-- ADDED: Base Model Filter Dropdown -->
-                                <select id="civitai-search-base-model" class="civitai-select">
-                                    <option value="any">Any Base Model</option>
-                                    <!-- Base models populated here -->
-                                </select>
-                                <select id="civitai-search-sort" class="civitai-select">
-                                    <option value="Relevancy">Relevancy</option>
-                                    <option value="Highest Rated">Highest Rated</option>
-                                    <option value="Most Liked">Most Liked</option>
-                                    <option value="Most Discussed">Most Discussed</option> 
-                                    <option value="Most Collected">Most Collected</option>
-                                    <option value="Most Buzz">Most Buzz</option>
-                                    <option value="Most Downloaded">Most Downloaded</option>
-                                    <option value="Newest">Newest</option>
-                                </select>
-                            </div>
-                        <button type="submit" id="civitai-search-submit" class="civitai-button primary">Search</button>
-                        </form>
-                        <div id="civitai-search-results" class="civitai-search-results">
-                            <!-- Search results -->
-                        </div>
-                        <div id="civitai-search-pagination" style="text-align: center; margin-top: 20px;">
-                            <!-- Pagination -->
-                        </div>
-                    </div>
-                    <div id="civitai-tab-status" class="civitai-downloader-tab-content">
-                        <div id="civitai-status-content">
-                             <div class="civitai-status-section">
-                                <h3>Active Downloads</h3>
-                                <div id="civitai-active-list" class="civitai-download-list">
-                                    <p>No active downloads.</p>
-                                </div>
-                            </div>
-                             <div class="civitai-status-section">
-                                <h3>Queued Downloads</h3>
-                                <div id="civitai-queued-list" class="civitai-download-list">
-                                    <p>Download queue is empty.</p>
-                                </div>
-                            </div>
-                             <div class="civitai-status-section">
-                                <div style="display: flex; justify-content: space-between; align-items: center; margin-bottom: 10px;">
-                                     <h3>Download History (Recent)</h3>
-                                     <button id="civitai-clear-history-button" class="civitai-button danger small" title="Clear all history items">
-                                         <i class="fas fa-trash-alt"></i> Clear History
-                                     </button>
-                                </div>
-                                <div id="civitai-history-list" class="civitai-download-list">
-                                    <p>No download history yet.</p>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                    <div id="civitai-tab-settings" class="civitai-downloader-tab-content">
-                        <form id="civitai-settings-form">
-                         <div class="civitai-settings-container">
-                             <div class="civitai-settings-section">
-                                 <h4>API & Defaults</h4>
-                                <div class="civitai-form-group">
-                                    <label for="civitai-settings-api-key">Civitai API Key (Optional)</label>
-                                    <input type="password" id="civitai-settings-api-key" class="civitai-input" placeholder="Enter API key for higher limits / authenticated access" autocomplete="new-password">
-                                    <p style="font-size: 0.85em; color: #bbb; margin-top: 5px;">Needed for some downloads/features. Find keys at civitai.com/user/account</p>
-                                </div>
-                                <div class="civitai-form-group">
-                                    <label for="civitai-settings-connections">Default Connections</label>
-                                    <input type="number" id="civitai-settings-connections" class="civitai-input" value="1" min="1" max="16" step="1" required disabled>
-                                    <p style="font-size: 0.85em; color: #bbb; margin-top: 5px;">Disabled. Only single connection possible for now</p>
-                                </div>
-                                <div class="civitai-form-group">
-                                    <label for="civitai-settings-default-type">Default Model Type (for saving)</label>
-                                    <select id="civitai-settings-default-type" class="civitai-select" required></select>
-                                </div>
-                             </div>
-                            <div class="civitai-settings-section">
-                                <h4>Interface & Search</h4>
-                                <div class="civitai-form-group inline">
-                                    <input type="checkbox" id="civitai-settings-auto-open-status" class="civitai-checkbox">
-                                    <label for="civitai-settings-auto-open-status">Switch to Status tab after starting download</label>
-                                </div>
-                                <!-- Example: Add search result limit setting -->
-                                <!--
-                                <div class="civitai-form-group">
-                                    <label for="civitai-settings-search-limit">Search Results per Page</label>
-                                    <input type="number" id="civitai-settings-search-limit" class="civitai-input" value="10" min="5" max="50" step="5" required>
-                                </div>
-                                -->
-                            </div>
-                         </div>
-                         <button type="submit" id="civitai-settings-save" class="civitai-button primary" style="margin-top: 20px;">Save Settings</button>
-                        </form>
-                    </div>
-                </div>
-                <!-- Toast Notification Area -->
-                <div id="civitai-toast" class="civitai-toast"></div>
-
-                 <!-- ADDED Confirmation Modal (hidden by default) -->
-                 <div id="civitai-confirm-clear-modal" class="civitai-confirmation-modal">
-                     <div class="civitai-confirmation-modal-content">
-                         <h4>Confirm Clear History</h4>
-                         <p>Are you sure you want to clear the download history? This action cannot be undone.</p>
-                         <div class="civitai-confirmation-modal-actions">
-                              <button id="civitai-confirm-clear-no" class="civitai-button secondary">Cancel</button>
-                              <button id="civitai-confirm-clear-yes" class="civitai-button danger">Confirm Clear</button>
-                         </div>
-                     </div>
-                 </div>
-
-            </div> <!-- End civitai-downloader-modal-content -->
-        `;
+        // Use the imported template function to populate the modal's HTML
+        this.modal.innerHTML = modalTemplate(this.settings);
     }
 
      cacheDOMElements() {
@@ -1247,206 +1080,7 @@ class CivitaiDownloaderUI {
 
     renderSearchResults(items) { // Expects the array of hits (`processed_items` from backend)
         return renderSearchResultsExternal(this, items);
-
-        const placeholder = PLACEHOLDER_IMAGE_URL; // Ensure this global var is accessible
-        const onErrorScript = `this.onerror=null; this.src='${placeholder}'; this.style.backgroundColor='#444';`;
-        const fragment = document.createDocumentFragment();
-
-        // console.log("Rendering Meili results:", items); // Debug: Log items received
-
-        items.forEach(hit => {
-            // --- Safely extract data from Meili hit object ---
-            const modelId = hit.id;
-            if (!modelId) {
-                console.warn("Skipping search result with missing ID:", hit);
-                return; // Skip if essential ID is missing
-            }
-            const creator = hit.user?.username || 'Unknown Creator';
-            const modelName = hit.name || 'Untitled Model';
-            const modelTypeApi = hit.type || 'N/A'; // API Type (e.g., LORA, Checkpoint)
-            const stats = hit.metrics || {}; // Top-level metrics
-            const tags = hit.tags?.map(t => t.name) || []; // Extract just the tag names
-
-            // Thumbnail URL (pre-processed by backend)
-            const thumbnailUrl = hit.thumbnailUrl || placeholder;
-            const thumbnailType = hit.images[0].type
-
-            // --- Version Info ---
-            const allVersions = hit.versions || []; // Array of all available versions
-            const primaryVersion = hit.version || (allVersions.length > 0 ? allVersions[0] : {}); // Primary version details provided directly, fallback to first in array
-            const primaryVersionId = primaryVersion.id;
-            const primaryBaseModel = primaryVersion.baseModel || 'N/A'; // Base model from primary version
-
-            // Get unique base models across *all* versions for display (more comprehensive)
-            const uniqueBaseModels = allVersions.length > 0
-                ? [...new Set(allVersions.map(v => v.baseModel).filter(Boolean))]
-                : (primaryBaseModel !== 'N/A' ? [primaryBaseModel] : []); // Fallback to primary if array empty
-            const baseModelsDisplay = uniqueBaseModels.length > 0 ? uniqueBaseModels.join(', ') : 'N/A';
-
-            // --- Latest Update Date ---
-            let lastUpdatedFormatted = 'N/A';
-            const publishedAt = hit.publishedAt; // Use publishedAt from main hit
-            if (publishedAt) {
-                try {
-                    const date = new Date(publishedAt);
-                    // Use a more standard locale format
-                    lastUpdatedFormatted = date.toLocaleDateString(undefined, { // undefined uses user's locale
-                        year: 'numeric', month: 'short', day: 'numeric'
-                    });
-                } catch (e) {
-                    console.error(`Error parsing date for model ${modelId}:`, publishedAt, e);
-                    lastUpdatedFormatted = 'Invalid Date';
-                }
-            }
-
-            // --- Create List Item ---
-            const listItem = document.createElement('div');
-            listItem.className = 'civitai-search-item';
-            listItem.dataset.modelId = modelId; // Add model ID for potential future use
-
-            // --- Generate Version Buttons ---
-            // Show primary version first, then others, up to a limit initially
-            const MAX_VISIBLE_VERSIONS = 3;
-            let visibleVersions = [];
-            if (primaryVersionId) {
-                 // Ensure primary version object from `hit.version` has necessary fields
-                 const primaryVersionData = {
-                     id: primaryVersionId,
-                     name: primaryVersion.name || 'Primary Version',
-                     baseModel: primaryBaseModel,
-                     // Add other fields if needed by the button template
-                 };
-                 visibleVersions.push(primaryVersionData);
-            }
-            // Add other versions, ensuring no duplicates with the primary one if it was also in the array
-            allVersions.forEach(v => {
-                 if (v.id !== primaryVersionId && visibleVersions.length < MAX_VISIBLE_VERSIONS) {
-                     visibleVersions.push(v);
-                 }
-            });
-
-            // Generate HTML for initially visible version buttons
-            let versionButtonsHtml = visibleVersions.map(version => {
-                const versionId = version.id;
-                const versionName = version.name || 'Unknown Version';
-                const baseModel = version.baseModel || 'N/A';
-                return `
-                    <button class="civitai-button primary small civitai-search-download-button"
-                            data-model-id="${modelId}"
-                            data-version-id="${versionId || ''}"
-                            data-model-type="${modelTypeApi || ''}"
-                            ${!versionId ? 'disabled title="Version ID missing, cannot pre-fill"' : 'title="Pre-fill Download Tab"'} >
-                        <span class="base-model-badge">${baseModel}</span> ${versionName} <i class="fas fa-download"></i>
-                    </button>
-                `;
-            }).join(''); // Join the HTML strings
-
-            // --- "All/More Versions" Button Logic ---
-            const hasMoreVersions = allVersions.length > visibleVersions.length;
-            const totalVersionCount = allVersions.length;
-            const moreButtonHtml = hasMoreVersions ? `
-                <button class="civitai-button secondary small show-all-versions-button"
-                        data-model-id="${modelId}"
-                        data-total-versions="${totalVersionCount}"
-                        title="Show all ${totalVersionCount} versions">
-                    All versions (${totalVersionCount}) <i class="fas fa-chevron-down"></i>
-                </button>
-            ` : '';
-
-            // --- Hidden Versions Container ---
-            let allVersionsHtml = '';
-            if (hasMoreVersions) {
-                // Get versions that are not already visible
-                const hiddenVersions = allVersions.filter(v => !visibleVersions.some(vis => vis.id === v.id));
-                allVersionsHtml = `
-                    <div class="all-versions-container" id="all-versions-${modelId}" style="display: none;">
-                        ${hiddenVersions.map(version => {
-                            const versionId = version.id;
-                            const versionName = version.name || 'Unknown Version';
-                            const baseModel = version.baseModel || 'N/A';
-                            return `
-                                <button class="civitai-button primary small civitai-search-download-button"
-                                        data-model-id="${modelId}"
-                                        data-version-id="${versionId || ''}"
-                                        data-model-type="${modelTypeApi || ''}"
-                                        ${!versionId ? 'disabled title="Version ID missing, cannot pre-fill"' : 'title="Pre-fill Download Tab"'} >
-                                    <span class="base-model-badge">${baseModel}</span> ${versionName} <i class="fas fa-download"></i>
-                                </button>
-                            `;
-                        }).join('')}
-                    </div>
-                `;
-            }
-
-            let thumbnailHtml = '';
-            const videoTitle = `Video preview for ${modelName}`;
-            const imageAlt = `${modelName} thumbnail`;
-
-            if (thumbnailUrl && typeof thumbnailUrl === 'string' && thumbnailType == "video") {
-                thumbnailHtml = `
-                    <video class="civitai-search-thumbnail"
-                           src="${thumbnailUrl}"
-                           autoplay loop muted playsinline
-                           title="${videoTitle}"
-                           onerror="console.error('Failed to load video preview:', this.src)">
-                           <!-- Optional: Fallback message if video tag not supported -->
-                           Your browser does not support the video tag.
-                    </video>
-                `;
-            } else {
-                // It's likely an IMAGE or missing, render IMG tag with fallback
-                const effectiveThumbnailUrl = thumbnailUrl || placeholder; // Use placeholder if URL is missing
-                thumbnailHtml = `
-                    <img src="${effectiveThumbnailUrl}"
-                         alt="${imageAlt}"
-                         class="civitai-search-thumbnail"
-                         loading="lazy"
-                         onerror="${onErrorScript}">
-                `;
-            }
-
-            // --- Construct Final Inner HTML for the List Item ---
-            listItem.innerHTML = `
-                <div class="civitai-thumbnail-container">
-                    ${thumbnailHtml}
-                </div>
-                <div class="civitai-search-info">
-                    <h4>${modelName}</h4>
-                    <div class="civitai-search-meta-info">
-                        <span title="Creator: ${creator}"><i class="fas fa-user"></i> ${creator}</span>
-                        <span title="Base Models: ${baseModelsDisplay}"><i class="fas fa-layer-group"></i> ${baseModelsDisplay}</span>
-                        <span title="Published: ${lastUpdatedFormatted}"><i class="fas fa-calendar-alt"></i> ${lastUpdatedFormatted}</span>
-                    </div>
-                    <div class="civitai-search-stats" title="Stats: Downloads / Rating (Count) / Likes">
-                        <span title="Downloads"><i class="fas fa-download"></i> ${stats.downloadCount?.toLocaleString() || 0}</span>
-                        <span title="Thumbs"><i class="fas fa-thumbs-up"></i> ${stats.thumbsUpCount?.toLocaleString() || 0}</span>
-                        <span title="Collected"><i class="fas fa-archive"></i> ${stats.collectedCount?.toLocaleString() || 0}</span>
-                        <span title="Buzz"><i class="fas fa-bolt"></i> ${stats.tippedAmountCount?.toLocaleString() || 0}</span>
-
-                    </div>
-                    ${tags.length > 0 ? `
-                    <div class="civitai-search-tags" title="${tags.join(', ')}">
-                        ${tags.slice(0, 5).map(tag => `<span class="civitai-search-tag">${tag}</span>`).join('')}
-                        ${tags.length > 5 ? `<span class="civitai-search-tag">...</span>` : ''}
-                    </div>
-                    ` : ''}
-                </div>
-                <div class="civitai-search-actions">
-                    <a href="https://civitai.com/models/${modelId}${primaryVersionId ? '?modelVersionId='+primaryVersionId : ''}" target="_blank" rel="noopener noreferrer" class="civitai-button small" title="Open on Civitai website">
-                        View <i class="fas fa-external-link-alt"></i>
-                    </a>
-                    <div class="version-buttons-container">
-                        ${versionButtonsHtml}
-                        ${moreButtonHtml} 
-                    </div>
-                    ${allVersionsHtml} 
-                </div> 
-            `; // End of listItem.innerHTML
-
-            fragment.appendChild(listItem); // Add the completed item to the fragment
-        }); // End forEach loop for hits
-
-    } // End renderSearchResults method
+    }
     
 
     renderSearchPagination(metadata) { // Expects the metadata object from backend { totalItems, currentPage, pageSize, totalPages, ... }
